refactor(sdk): dedup operation-discovery test helper
client-facade-completeness.test.ts and input-schema-strict-guard.test.ts
each copy-pasted the same import.meta.glob-based Operation discovery
(isOperation/loadAllOperations plus the ImportMeta.glob ambient
declaration). Extracted into a shared __tests__/support/load-operations.ts
so the two structural guard tests can't drift on what counts as an
Operation.

Surfaced by an independent code-review pass on PR #1845 (reuse +
simplification angles, both corroborating this finding; CONFIRMED on
verification).
